HOTT-2643 Refactor Heading serializers
Cleans up code to make extending with Nested Set support easier
